Add proper min/max settings to pedal assign effect.
1) Delete definition of pedal1_assign_settings since it should be calculated
   at runtime from the modifier list.

2) Delete the pedal1_assign_settings EffectSetting from
rp355_pedal1_assign_group.

3) Add "None" entries to the tone_lib_group[] and effect_lib_group[]. Otherwise,
   once a library tone or effect is selected, it can't be deselected.

4) Fix up some tracing of effect values that was displaying incorrect values.

5) Change 'update_modifier_combo_box()' to 'update_modifier_vbox()' and for
   the expression pedal let it handle instantiating per combo box entry
   settings.

6) Don't forget to free the links we remove from the list when we're cleaning
   the combo box. If there are per-combo box entry settings, free them using
   the usual destructor.

7) Poll the device for the appropriate settings for EXP_MIN/MAX after we
   we update the modifier list.

+++ gdigi.c
@@ -154,10 +154,10 @@ format_value (XmlSettings *xml, guint value)
     case VALUE_TYPE_PLAIN:
     {
         if (decimal) {
-            double dvalue = (value + offset) * step;
+            double dvalue = ((gint)value + offset) * step;
                 g_string_printf(buf, "%0.2f%s", dvalue, suffix);
         } else {
-            gint ivalue = (value + offset) * step;
+            gint ivalue = ((gint)value + offset) * step;
             g_string_printf(buf, "%d%s", ivalue, suffix);
         }
         break;
